+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## Unreleased
 
 ### Improvements
+
+- fix(cluster): fix name ref error when using instanceProfile
+  [#189](https://github.com/pulumi/pulumi-eks/pull/189)
 - fix(cluster): support configuring private and public endpoint access
   [#154](https://github.com/pulumi/pulumi-eks/pull/154)
 - fix(cluster): support passing additional arguments to /etc/eks/bootstrap.sh and --kubelet-extra-args
   [#181](https://github.com/pulumi/pulumi-eks/pull/181)
-
 
 ## 0.18.8 (Released June 19, 2019)