+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## Unreleased
 
 ### Improvements
+
+- fix(aws-kubectl): add retry logic to the kubectl command for CNI & aws-auth
+  [#186](https://github.com/pulumi/pulumi-eks/pull/186)
 - fix(cluster): support configuring private and public endpoint access
   [#154](https://github.com/pulumi/pulumi-eks/pull/154)
 - fix(cluster): support passing additional arguments to /etc/eks/bootstrap.sh and --kubelet-extra-args
   [#181](https://github.com/pulumi/pulumi-eks/pull/181)
-
 
 ## 0.18.8 (Released June 19, 2019)